requirements.txt
streamlit>=1.30.0
groq>=0.4.0
pandas>=2.0.0
plotly>=5.18.0
sqlparse>=0.5.0
faker>=22.0.0
python-dotenv>=1.0.0
